different scripts for 1) mpi4py creation of derived fields, one table per brick 2) merging those tables into one table for all brick 3) making plots using the merged table
