+++ py/obiwan/qa/heatmaps.py
@@ -532,328 +532,6 @@ def plots(opt):
         
     return 0
        
-def _ring_unique(wcs, W, H, i, unique, ra1,ra2,dec1,dec2):
-    lo, hix, hiy = i, W-i-1, H-i-1
-    # one slice per side; we double-count the last pix of each side.
-    sidex = slice(lo,hix+1)
-    sidey = slice(lo,hiy+1)
-    top = (lo, sidex)
-    bot = (hiy, sidex)
-    left  = (sidey, lo)
-    right = (sidey, hix)
-    xx = np.arange(W)
-    yy = np.arange(H)
-    nu,ntot = 0,0
-    for slc in [top, bot, left, right]:
-        #print('xx,yy', xx[slc], yy[slc])
-        (yslc,xslc) = slc
-        rr,dd = wcs.pixelxy2radec(xx[xslc]+1, yy[yslc]+1)
-        U = (rr >= ra1 ) * (rr < ra2 ) * (dd >= dec1) * (dd < dec2)
-        #print('Pixel', i, ':', np.sum(U), 'of', len(U), 'pixels are unique')
-        unique[slc] = U
-        nu += np.sum(U)
-        ntot += len(U)
-    #if allin:
-    #    print('Scanned to pixel', i)
-    #    break
-    return nu,ntot
-
-def find_unique_pixels(wcs, W, H, unique, ra1,ra2,dec1,dec2):
-    if unique is None:
-        unique = np.ones((H,W), bool)
-    # scan the outer annulus of pixels, and shrink in until all pixels
-    # are unique.
-    step = 10
-    for i in range(0, W//2, step):
-        nu,ntot = _ring_unique(wcs, W, H, i, unique, ra1,ra2,dec1,dec2)
-        #print('Pixel', i, ': nu/ntot', nu, ntot)
-        if nu > 0:
-            i -= step
-            break
-        unique[:i,:] = False
-        unique[H-1-i:,:] = False
-        unique[:,:i] = False
-        unique[:,W-1-i:] = False
-
-    for j in range(max(i+1, 0), W//2):
-        nu,ntot = _ring_unique(wcs, W, H, j, unique, ra1,ra2,dec1,dec2)
-        #print('Pixel', j, ': nu/ntot', nu, ntot)
-        if nu == ntot:
-            break
-    return unique
-
-
-
-def merge_tables_divide(bricklist,nproc,outfn)
-    bricks = fits_table(os.path.join(os.environ['LEGACY_SURVEY_DIR'],
-                                     'survey-bricks.fits.gz'))
-    if nproc > 1:
-        from mpi4py.MPI import COMM_WORLD as comm
-        bricklist= np.array_split(bricklist, comm.size)[comm.rank]
-    else:
-        class MyComm(object):
-            def __init__(self):
-                self.rank=0
-        comm= MyComm()
-
-    one_row_per_brick= []
-    one_row_per_brick.append( mpi_merge_conquer(bricklist) )
-    one_row_per_brick = comm.gather(one_row_per_brick, root=0)
-    if comm.rank == 0:
-        allRows = merge_tables(one_row_per_brick)
-        allRows.writeto(outfn)
-        print('Wrote', outfn)
-
-
-def mpi_merge_conquer(bricklist)
-    brickset = set()
-    gn = []
-    rn = []
-    zn = []
-    
-    gnhist = []
-    rnhist = []
-    znhist = []
-    
-    nnhist = 6
-    
-    gdepth = []
-    rdepth = []
-    zdepth = []
-    
-    ibricks = []
-    nsrcs = []
-    npsf  = []
-    nsimp = []
-    nrex = []
-    nexp  = []
-    ndev  = []
-    ncomp = []
-
-    gpsfsize = []
-    rpsfsize = []
-    zpsfsize = []
-
-    gpsfdepth = []
-    rpsfdepth = []
-    zpsfdepth = []
-    ggaldepth = []
-    rgaldepth = []
-    zgaldepth = []
-
-    wise_nobs = []
-    wise_trans = []
-    
-    ebv = []
-    gtrans = []
-    rtrans = []
-    ztrans = []
-    
-    
-    #sfd = SFDMap()
-    
-    W = H = 3600
-    # H=3600
-    # xx,yy = np.meshgrid(np.arange(W), np.arange(H))
-    unique = np.ones((H,W), bool)
-    tlast = 0
-    
-    for ibrick,brick in enumerate(bricklist):
-        if ibrick % 10 == 0: 
-            print('rank ',comm.rank,': File', (ibrick+1), 'of', len(bricklist), ':', brick)
-        #words = fn.split('/')
-        #dirprefix = '/'.join(words[:-4])
-        #print('Directory prefix:', dirprefix)
-        #words = words[-4:]
-        #brick = words[2]
-        #print('Brick', brick)
-        try:
-            tfn = os.path.join(dirprefix, 'tractor', brick[:3], 'tractor-%s.fits'%brick)
-            print('Tractor filename', tfn)
-            if opt.dr5:
-                T = fits_table(tfn, columns=['brick_primary', 'type',
-                                             'psfsize_g', 'psfsize_r', 'psfsize_z',
-                                             'psfdepth_g', 'psfdepth_r', 'psfdepth_z',
-                                             'galdepth_g', 'galdepth_r', 'galdepth_z',
-                                             'ebv',
-                                             'mw_transmission_g', 'mw_transmission_r', 'mw_transmission_z',
-                                             'nobs_w1', 'nobs_w2', 'nobs_w3', 'nobs_w4',
-                                             'mw_transmission_w1', 'mw_transmission_w2', 'mw_transmission_w3', 'mw_transmission_w4'])
-            else:
-                T = fits_table(tfn, columns=['brick_primary', 'type', 'decam_psfsize',
-                                         'decam_depth', 'decam_galdepth',
-                                         'ebv', 'decam_mw_transmission',
-                                         'wise_nobs', 'wise_mw_transmission'])
-        except:
-            print('Failed to read FITS table', tfn)
-            import traceback
-            traceback.print_exc()
-            print('Carrying on.')
-            continue
-
-        brickset.add(brick)
-        bricklist.append(brick)
-        gn.append(0)
-        rn.append(0)
-        zn.append(0)
-
-        gnhist.append([0 for i in range(nnhist)])
-        rnhist.append([0 for i in range(nnhist)])
-        znhist.append([0 for i in range(nnhist)])
-
-        index = -1
-        ibrick = np.nonzero(bricks.brickname == brick)[0][0]
-        ibricks.append(ibrick)
-
-        T.cut(T.brick_primary)
-        nsrcs.append(len(T))
-        types = Counter([t.strip() for t in T.type])
-        npsf.append(types['PSF'])
-        nsimp.append(types['SIMP'])
-        nrex.append(types['REX'])
-        nexp.append(types['EXP'])
-        ndev.append(types['DEV'])
-        ncomp.append(types['COMP'])
-        print('N sources', nsrcs[-1])
-
-        if opt.dr5:
-            gpsfsize.append(np.median(T.psfsize_g))
-            rpsfsize.append(np.median(T.psfsize_r))
-            zpsfsize.append(np.median(T.psfsize_z))
-
-            gpsfdepth.append(np.median(T.psfdepth_g))
-            rpsfdepth.append(np.median(T.psfdepth_r))
-            zpsfdepth.append(np.median(T.psfdepth_z))
-
-            ggaldepth.append(np.median(T.galdepth_g))
-            rgaldepth.append(np.median(T.galdepth_r))
-            zgaldepth.append(np.median(T.galdepth_z))
-
-            wise_nobs.append(np.median(
-                np.vstack((T.nobs_w1, T.nobs_w2, T.nobs_w3, T.nobs_w4)).T,
-                axis=0))
-            wise_trans.append(np.median(
-                np.vstack((T.mw_transmission_w1,
-                           T.mw_transmission_w2,
-                           T.mw_transmission_w3,
-                           T.mw_transmission_w4)).T,
-                           axis=0))
-
-            gtrans.append(np.median(T.mw_transmission_g))
-            rtrans.append(np.median(T.mw_transmission_r))
-            ztrans.append(np.median(T.mw_transmission_z))
-            
-        else:
-            gpsfsize.append(np.median(T.decam_psfsize[:,1]))
-            rpsfsize.append(np.median(T.decam_psfsize[:,2]))
-            zpsfsize.append(np.median(T.decam_psfsize[:,4]))
-
-            gpsfdepth.append(np.median(T.decam_depth[:,1]))
-            rpsfdepth.append(np.median(T.decam_depth[:,2]))
-            zpsfdepth.append(np.median(T.decam_depth[:,4]))
-
-            ggaldepth.append(np.median(T.decam_galdepth[:,1]))
-            rgaldepth.append(np.median(T.decam_galdepth[:,2]))
-            zgaldepth.append(np.median(T.decam_galdepth[:,4]))
-
-            wise_nobs.append(np.median(T.wise_nobs, axis=0))
-            wise_trans.append(np.median(T.wise_mw_transmission, axis=0))
-
-            gtrans.append(np.median(T.decam_mw_transmission[:,1]))
-            rtrans.append(np.median(T.decam_mw_transmission[:,2]))
-            ztrans.append(np.median(T.decam_mw_transmission[:,4]))
-            
-        ebv.append(np.median(T.ebv))
-
-        br = bricks[ibrick]
-
-        #print('Computing unique brick pixels...')
-        pixscale = 0.262/3600.
-        wcs = Tan(br.ra, br.dec, W/2.+0.5, H/2.+0.5,
-                  -pixscale, 0., 0., pixscale,
-                  float(W), float(H))
-        unique[:,:] = True
-        find_unique_pixels(wcs, W, H, unique,
-                           br.ra1, br.ra2, br.dec1, br.dec2)
-        U = np.flatnonzero(unique)
-        #print(len(U), 'of', W*H, 'pixels are unique to this brick')
-         
-        index = bricklist.index(brick)
-        assert(index == len(bricklist)-1)
-    
-        filepart = words[-1]
-        filepart = filepart.replace('.fits.gz', '')
-        filepart = filepart.replace('.fits.fz', '')
-        print('File:', filepart)
-        band = filepart[-1]
-        assert(band in 'grz')
-    
-        nlist,nhist = dict(g=(gn,gnhist), r=(rn,rnhist), z=(zn,znhist))[band]
-    
-        upix = fitsio.read(fn).flat[U]
-        med = np.median(upix)
-        print('Band', band, ': Median', med)
-        nlist[index] = med
-    
-        hist = nhist[index]
-        for i in range(nnhist):
-            if i < nnhist-1:
-                hist[i] = np.sum(upix == i)
-            else:
-                hist[i] = np.sum(upix >= i)
-        assert(sum(hist) == len(upix))
-        print('Number of exposures histogram:', hist)
-    
-    ibricks = np.array(ibricks)
-    
-    #print('Maximum number of sources:', max(nsrcs))
-    
-    T = fits_table()
-    T.brickname = np.array(bricklist)
-    T.ra  = bricks.ra [ibricks]
-    T.dec = bricks.dec[ibricks]
-    T.nexp_g = np.array(gn).astype(np.int16)
-    T.nexp_r = np.array(rn).astype(np.int16)
-    T.nexp_z = np.array(zn).astype(np.int16)
-    T.nexphist_g = np.array(gnhist).astype(np.int32)
-    T.nexphist_r = np.array(rnhist).astype(np.int32)
-    T.nexphist_z = np.array(znhist).astype(np.int32)
-    T.nobjs  = np.array(nsrcs).astype(np.int16)
-    T.npsf   = np.array(npsf ).astype(np.int16)
-    T.nsimp  = np.array(nsimp).astype(np.int16)
-    T.nrex   = np.array(nrex ).astype(np.int16)
-    T.nexp   = np.array(nexp ).astype(np.int16)
-    T.ndev   = np.array(ndev ).astype(np.int16)
-    T.ncomp  = np.array(ncomp).astype(np.int16)
-    T.psfsize_g = np.array(gpsfsize).astype(np.float32)
-    T.psfsize_r = np.array(rpsfsize).astype(np.float32)
-    T.psfsize_z = np.array(zpsfsize).astype(np.float32)
-    with np.errstate(divide='ignore'):
-        T.psfdepth_g = (-2.5*(-9.+np.log10(5.*np.sqrt(1. / np.array(gpsfdepth))))).astype(np.float32)
-        T.psfdepth_r = (-2.5*(-9.+np.log10(5.*np.sqrt(1. / np.array(rpsfdepth))))).astype(np.float32)
-        T.psfdepth_z = (-2.5*(-9.+np.log10(5.*np.sqrt(1. / np.array(zpsfdepth))))).astype(np.float32)
-        T.galdepth_g = (-2.5*(-9.+np.log10(5.*np.sqrt(1. / np.array(ggaldepth))))).astype(np.float32)
-        T.galdepth_r = (-2.5*(-9.+np.log10(5.*np.sqrt(1. / np.array(rgaldepth))))).astype(np.float32)
-        T.galdepth_z = (-2.5*(-9.+np.log10(5.*np.sqrt(1. / np.array(zgaldepth))))).astype(np.float32)
-    for k in ['psfdepth_g', 'psfdepth_r', 'psfdepth_z', 'galdepth_g', 'galdepth_r', 'galdepth_z']:
-        v = T.get(k)
-        v[np.logical_not(np.isfinite(v))] = 0.
-    T.ebv = np.array(ebv).astype(np.float32)
-    T.trans_g = np.array(gtrans).astype(np.float32)
-    T.trans_r = np.array(rtrans).astype(np.float32)
-    T.trans_z = np.array(ztrans).astype(np.float32)
-    T.ext_g = -2.5 * np.log10(T.trans_g)
-    T.ext_r = -2.5 * np.log10(T.trans_r)
-    T.ext_z = -2.5 * np.log10(T.trans_z)
-    T.wise_nobs = np.array(wise_nobs).astype(np.int16)
-    T.trans_wise = np.array(wise_trans).astype(np.float32)
-    T.ext_w1 = -2.5 * np.log10(T.trans_wise[:,0])
-    T.ext_w2 = -2.5 * np.log10(T.trans_wise[:,1])
-    T.ext_w3 = -2.5 * np.log10(T.trans_wise[:,2])
-    T.ext_w4 = -2.5 * np.log10(T.trans_wise[:,3])
-
-    return T 
-
 
 
 def main():
